requirements.txt
websocket-client
pyspark
finnhub-python
python-dotenv
redis
sqlalchemy